start working on PlayerDocumentConverterTest
