+++ app/src/main/java/com/example/hashcache/javadocs/com/example/hashcache/tests/SearchTest.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.hashcache.tests.SearchTest</li>
+<li>com.example.hashcache.tests.PlayerListTest</li>
 </ul>
 </li>
 </ul>
