+++ app/src/test/java/com/example/hashcache/tests/PlayerListTest.java
@@ -1,23 +1,9 @@
 package com.example.hashcache.tests;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import com.example.hashcache.models.database.DatabaseAdapters.callbacks.GetPlayerCallback;
-import com.example.hashcache.models.database.DatabaseAdapters.PlayersDatabaseAdapter;
 import com.example.hashcache.models.Player;
 import com.example.hashcache.models.PlayerList;
+import com.example.hashcache.models.database.DatabaseAdapters.PlayersDatabaseAdapter;
+import com.example.hashcache.models.database.DatabaseAdapters.callbacks.BooleanCallback;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,61 +12,43 @@ import org.mockito.Mockito;
 import java.util.ArrayList;
 
 public class PlayerListTest {
+
     private String newPlayerUsername;
-    private Player mockPlayer;
-    private ArrayList<String> names;
     private PlayersDatabaseAdapter mockPlayerConnectionHandler;
-    private GetStringCallback mockGetStringCallback;
+    private BooleanCallback mockBooleanCallback;
 
     @BeforeEach
     void resetMocks(){
         newPlayerUsername = "Stubby";
-        mockPlayer = new Player(newPlayerUsername);
-        names = new ArrayList<>();
         mockPlayerConnectionHandler = Mockito.mock(PlayersDatabaseAdapter.class);
-        mockGetStringCallback = Mockito.mock(GetStringCallback.class);
+        mockBooleanCallback = Mockito.mock(BooleanCallback.class);
         PlayerList.resetInstance();
     }
 
     @Test
-    void testAddPlayerSuccess(){
-        GetStringCallback mockGetStringCallback = Mockito.mock(GetStringCallback.class);
-        Player mockPlayer = new Player("name");
-
-        when(mockPlayerConnectionHandler.getInAppPlayerUserNames()).thenReturn(names);
-        doAnswer(invocation -> {
-            names.add(newPlayerUsername);
-            mockGetStringCallback.onCallback("id");
-            return null;
-        }).when(mockPlayerConnectionHandler).createPlayer(mockPlayer.getUsername(), mockGetStringCallback);
-
+    public void testLevenshtein() {
         PlayerList playerList = PlayerList.getInstance(mockPlayerConnectionHandler);
-        playerList.addPlayer(mockPlayer.getUsername(), mockGetStringCallback);
+        Integer distance = playerList.computeLevenshteinDistance("benyam", "ephrem");
 
-        verify(mockPlayerConnectionHandler, times(1)).createPlayer(anyString(),
-                any(GetStringCallback.class));
+        assert distance == 5;
+
+        distance = playerList.computeLevenshteinDistance("hashcache", "hashcache");
+
+        assert distance == 0;
+
+        distance = playerList.computeLevenshteinDistance("a", "Ryan");
+
+        assert distance == 3;
     }
 
+    // TODO add a test for searching
     @Test
-    void testAddPlayerFailure(){
-        doThrow(new IllegalArgumentException()).when(mockPlayerConnectionHandler).createPlayer(
-                anyString(), any(GetStringCallback.class));
-
-        PlayerList playerList = PlayerList.getInstance(mockPlayerConnectionHandler);
-
-        assertFalse(playerList.addPlayer(newPlayerUsername, mockGetStringCallback));
+    public void testSearch() {
     }
 
+    // TODO add a test for using the filter. When more stuff with the leaderboard is setup
     @Test
-    void testGetPlayer(){
-        GetPlayerCallback mockGetPlayerCallback = Mockito.mock(GetPlayerCallback.class);
+    public void testFilter() {
 
-        doNothing().when(mockPlayerConnectionHandler).getPlayer(newPlayerUsername, mockGetPlayerCallback);
-
-        PlayerList playerList = PlayerList.getInstance(mockPlayerConnectionHandler);
-        playerList.getPlayer(newPlayerUsername, mockGetPlayerCallback);
-
-        verify(mockPlayerConnectionHandler, times(1))
-                .getPlayer(newPlayerUsername, mockGetPlayerCallback);
     }
 }
